frontend: fix code that symlinks external assets
We now check whether the necessary packages are installed, and abort if
not.

+++ squad/frontend/__init__.py
@@ -2,13 +2,23 @@ import os
 
 static = os.path.join(os.path.dirname(__file__), 'static')
 
-links = {
-    'angularjs': '/usr/share/javascript/angular.js',
-    'bootstrap': '/usr/share/javascript/bootstrap',
-    'font-awesome': '/usr/share/fonts-font-awesome',
-}
+links = [
+    ('angularjs', 'libjs-angularjs', '/usr/share/javascript/angular.js'),
+    ('bootstrap', 'libjs-bootstrap', '/usr/share/javascript/bootstrap'),
+    ('font-awesome', 'fonts-font-awesome', '/usr/share/fonts-font-awesome'),
+]
 
-for link, target in links.items():
-    link_path = os.path.join(static, link)
-    if not os.path.exists(link_path):
-        os.symlink(target, link_path)
+failed = False
+for lib, package, target in links:
+    link_path = os.path.join(static, lib)
+    if os.path.exists(target):
+        if not os.path.exists(link_path):
+            os.symlink(target, link_path)
+    else:
+        print("E: %s does not exist. Try installing %s" % (target, package))
+        print("I: You can also manually download %s to %s" % (lib, link_path))
+
+        failed = True
+
+if failed:
+    exit(1)
